Important changes to X86/ARM service runtime. Cleaning NaClSyscallCSegHook, getting rid of depreciated code and comments
Review URL: http://codereview.chromium.org/174009

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@572 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

+++ src/trusted/service_runtime/arch/arm/sel_rt.c
@@ -61,12 +61,12 @@ int NaClThreadContextCtor(struct NaClThreadContext  *ntcp,
 }
 
 
-uint32_t NaClGetThreadCtxSp(struct NaClThreadContext  *th_ctx) {
+uintptr_t NaClGetThreadCtxSp(struct NaClThreadContext  *th_ctx) {
   return (uintptr_t) th_ctx->stack_ptr;
 }
 
 
-void NaClSetThreadCtxSp(struct NaClThreadContext  *th_ctx, uint32_t sp) {
-  th_ctx->stack_ptr = sp;
+void NaClSetThreadCtxSp(struct NaClThreadContext  *th_ctx, uintptr_t sp) {
+  th_ctx->stack_ptr = (uint32_t) sp;
 }
